third_party/WebKit/LayoutTests/fast/dom/shadow/distribution-for-detached-subtree.html

<!DOCTYPE html>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
<script>
description('distribution should be recalculated correctly for a detached subtree.');

var shadowHost = document.createElement('div');
var shadowRoot = shadowHost.createShadowRoot();
var content = document.createElement('content');
shadowRoot.appendChild(content);
shouldBeEqualAsArray(content.getDistributedNodes(), []);

var shadowHostChild = document.createElement('div')
shadowHost.appendChild(shadowHostChild);
shouldBeEqualAsArray(content.getDistributedNodes(), [shadowHostChild]);
</script>
