third_party/WebKit/LayoutTests/fast/dom/shadow/distribution-for-detached-subtree-expected.txt

distribution should be recalculated correctly for a detached subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS successfullyParsed is true

TEST COMPLETE
